requirements.txt
py-clob-client>=0.20.0
httpx>=0.27.0
tenacity>=8.2.3
APScheduler>=3.10.4
python-dotenv>=1.0.1
pydantic>=2.7.0
